Idの型をint -> longに変更、次ページ存在判定が間違っていたところを修正

+++ Rialto/Models/DAO/Entity/ImageScore.cs
@@ -8,8 +8,8 @@ namespace Rialto.Models.DAO.Entity
 {
     public class ImageScore
     {
-        public int RegisterImageId { get; set; }
-        public int TagId { get; set; }
+        public long RegisterImageId { get; set; }
+        public long TagId { get; set; }
         public int Score { get; set; }
         public DateTime CreatedAt { get; set; }
         public DateTime UpdatedAt { get; set; }
