+++ Rialto/Models/DAO/Entity/TagAssign.cs
@@ -8,8 +8,8 @@ namespace Rialto.Models.DAO.Entity
 {
     public class TagAssign
     {
-        public int RegisterImageId { get; set; }
-        public int TagId { get; set; }
+        public long RegisterImageId { get; set; }
+        public long TagId { get; set; }
         public DateTime CreatedAt { get; set; }
         public DateTime UpdatedAt { get; set; }
     }
